+++ Changelog.md
@@ -1,3 +1,5 @@
 ## DART 4.0
 
 ### DART 4.0.0 (2014-0X-XX)
+
+1. Added optimizer using (`nlopt`)[http://ab-initio.mit.edu/wiki/index.php/NLopt]